2022-JUN-25
Thx, Peekobot! 
https://github.com/Peekobot/peekobot
https://twitter.com/magicroundabout/
	
My wife loves to solve quizzes. So I wrote an addon for the Peekobot so that the Peekobot can ask a quiz.

Usage: 
• Include the script with <script src="quiz_en.js"></script> in the index.html, before peekobot.js.
• Connect the quiz in conversations.js with the node "quiz_intro".
• fill out "your_exit_node" (see the variable below). Which this node you can connect the exiting of the script with your current conversation flow (for example to go back to the main menu)
• Look at the sample questions. This way you can add as many questions as you like.

The questions are shuffled, the answers are shuffled. A highscore is counted (by setting a cookie). At the end there is the possibility to display the highscore.

Have fun with the script!

License of the script: Public Domain / The Unlicense